requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
email-validator==2.1.0
WTForms==3.1.1
Flask-SocketIO==5.3.6
python-engineio==4.8.0
python-socketio==5.10.0

# Para PostgreSQL en producción (descomentar si lo usás)
# psycopg2-binary==2.9.9